Fix bug with missing revocations in JSON changelog
Some revocations (e.g. "T1053.001 At (Linux)" was revoked in ATT&CK
v11) appear in the Detailed HTML report but not in the JSON changelog.
This appears to be related to the way that changes are grouped in
DiffStix.get_groupings(). This commit fixes the missing revocations
in my testing.